try to mitigate none error for imagedownloader, fix later

+++ main.py
@@ -11,10 +11,15 @@ def submit_reddit_post_for_twitter():
     else:
         image_filename = reddit.download_post_image(post)
 
-    caption = f'"{post.title}"\n\nOn r/{post.subreddit.display_name} by u/{post.author.name}\n\n{twitter.get_hashtags_for_subreddit(post.subreddit.display_name)}'
-    twitter.upload_media_tweet(caption, image_filename)
-    os.remove(image_filename)
+    if image_filename is not None:
+        caption = f'"{post.title}"\n\nOn r/{post.subreddit.display_name} by u/{post.author.name}\n\n{twitter.get_hashtags_for_subreddit(post.subreddit.display_name)}'
+        twitter.upload_media_tweet(caption, image_filename)
+        os.remove(image_filename)
+    else:
+        submit_question_on_twitter() 
+        # eventually find a way to fix the none error, for now just give up
 
+       
 def submit_question_on_twitter():
     random = questions.get_random_question()
     hashtags = twitter.get_hashtags_for_question_type(random["type"])
@@ -22,6 +27,4 @@ def submit_question_on_twitter():
 
     twitter.upload_text_tweet(caption)
 
-
-
-submit_question_on_twitter()
+submit_reddit_post_for_twitter()
